requirements.txt
pandas
langchain_ollama
openai
python-dotenv
openpyxl
langchain
langchain_community